+++ OpenGLTester/source/BaseClasses/RenderObjects/Public/RenderObject.h
@@ -12,8 +12,8 @@ public:
     virtual void addVertex(Vertex&& _vertex) = 0;
     virtual std::vector<SPTR<Vertex>> getVertices() = 0;
     virtual size_t getVertexCount() const = 0;
-    virtual std::vector<unsigned int> getIndices() = 0;
-    virtual void adjustIndices(unsigned int _adjustment) = 0;
+    virtual std::vector<size_t> getIndices() = 0;
+    virtual void adjustIndices(size_t _adjustment) = 0;
 
     virtual uint32_t getRenderOrder() {return m_renderOrder;}
     virtual void setRenderOrder(uint32_t _newRo) {m_renderOrder = _newRo;}
